+++ src/main/kotlin/command/queryCommand.kt
@@ -1,5 +1,6 @@
 package cn.huohuas001.command
 
+import cn.huohuas001.tools.manager.BanManager
 import cn.huohuas001.tools.manager.ClientManager
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -13,9 +14,15 @@ object queryCommand: BaseCommand() {
         }
 
         val serverId = args[0]
-        val serverPackage = ClientManager.getServerPackageById(serverId);
-        if(serverPackage?.mServerClient == null){
-            logger.info("未查找到 $serverId 的服务器")
+        val serverPackage = ClientManager.getServerPackageById(serverId)
+
+        if(BanManager.queryBanReason(serverId) != null){
+            logger.info("服务器 $serverId 已被封禁,封禁理由为: ${BanManager.queryBanReason(serverId)}")
+            return false
+        }
+
+        if(serverPackage.mServerClient == null){
+            logger.warn("未查找到 $serverId 的服务器")
             return false
         }
         logger.info("服务器 $serverId 的详细为: ${serverPackage.mServerClient.name}|${serverPackage.mServerClient.version}|${serverPackage.mServerClient.platform}|${serverPackage.mServerClient.mSession.mIp}")